Completed API endpoints / API views

+++ sales/api/sales_rest/api_views.py
@@ -65,6 +65,14 @@ class SaleEncoder(ModelEncoder):
 
 @require_http_methods(["GET", "POST"])
 def api_list_salespeople(request):
+    try:
+        salespeople = Salesperson.objects.all()
+    except Salesperson.DoesNotExist:
+        return JsonResponse(
+            {"message": "Requested list of salespeople does not exist"},
+            status=404,
+        )
+
     if request.method == "GET":
         salespeople = Salesperson.objects.all()
         return JsonResponse(
@@ -83,6 +91,14 @@ def api_list_salespeople(request):
 
 @require_http_methods(["GET", "PUT", "DELETE"])
 def api_detail_salesperson(request, id):
+    try:
+        salesperson = Salesperson.objects.get(id=id)
+    except Salesperson.DoesNotExist:
+        return JsonResponse(
+            {"message": "Requested Salesperson does not exist"},
+            status=404,
+        )
+
     if request.method == "GET":
         salesperson = Salesperson.objects.get(id=id)
         return JsonResponse(
@@ -111,6 +127,14 @@ def api_detail_salesperson(request, id):
 
 @require_http_methods(["GET", "POST"])
 def api_list_customers(request):
+    try:
+        customers = Customer.objects.all()
+    except Customer.DoesNotExist:
+        return JsonResponse(
+            {"message": "Requested list of customers does not exist"},
+            status=404,
+        )
+
     if request.method == "GET":
         customers = Customer.objects.all()
         return JsonResponse(
@@ -129,6 +153,14 @@ def api_list_customers(request):
 
 @require_http_methods(["GET", "PUT", "DELETE"])
 def api_detail_customer(request, id):
+    try:
+        customer = Customer.objects.get(id=id)
+    except Customer.DoesNotExist:
+        return JsonResponse(
+            {"message": "Requested Customer does not exist"},
+            status=404,
+        )
+
     if request.method == "GET":
         customer = Customer.objects.get(id=id)
         return JsonResponse(
@@ -146,5 +178,92 @@ def api_detail_customer(request, id):
             safe=False,
         )
     else:  # DELETE
-        count, _ = Customer.objects.filter(id=id).delete()
+        customer = Customer.objects.get(id=id)
+        if not customer:
+            return JsonResponse(
+                {"message": "Requested Customer does not exist"},
+                status=404,
+            )
+        else:
+            count, _ = Customer.objects.filter(id=id).delete()
+            return JsonResponse({"deleted": count > 0})
+
+
+# ==============================
+
+# Sales:
+
+
+@require_http_methods(["GET", "POST"])
+def api_list_sales(request):
+    try:
+        sales = Sale.objects.all()
+    except Sale.DoesNotExist:
+        return JsonResponse(
+            {"message": "Requested list of sales does not exist"},
+            status=404,
+        )
+
+    if request.method == "GET":
+        sales = Sale.objects.all()
+        return JsonResponse(
+            {"sales": sales},
+            encoder=SaleEncoder,
+        )
+    else:  # POST
+        content = json.loads(request.body)
+
+        try:
+            content["automobile"] = AutomobileVO.objects.get(vin=content["automobile"])
+            content["salesperson"] = Salesperson.objects.get(id=content["salesperson"])
+            content["customer"] = Customer.objects.get(id=content["customer"])
+        except AutomobileVO.DoesNotExist:
+            return JsonResponse({"message": "Invalid VIN number"}, status=400)
+        except Salesperson.DoesNotExist:
+            return JsonResponse({"message": "Invalid Salesperson ID"}, status=400)
+        except Customer.DoesNotExist:
+            return JsonResponse({"message": "Invalid Customer ID"}, status=400)
+
+        new_sale = Sale.objects.create(**content)
+        return JsonResponse(
+            new_sale,
+            encoders=SaleEncoder,
+            safe=False,
+        )
+
+
+@require_http_methods(["GET", "PUT", "DELETE"])
+def api_detail_sale(request, id):
+    try:
+        sale = Sale.objects.get(id=id)
+    except Sale.DoesNotExist:
+        return JsonResponse({"message": "Requested Sale does not exist"}, status=404)
+
+    if request.method == "GET":
+        sale = Sale.objects.get(id=id)
+        return JsonResponse(
+            {"sale": sale},
+            encoders=SaleEncoder,
+            safe=False,
+        )
+    elif request.method == "PUT":
+        updated_content = json.loads(request.body)
+
+        try:
+            if "vin" in updated_content:
+                updated_content["automobile"] = AutomobileVO.objects.get(
+                    vin=updated_content["automobile"]
+                )
+        except AutomobileVO.DoesNotExist:
+            return JsonResponse({"message": "Invalid VIN number"}, status=400)
+
+        Sale.objects.filter(id=id).update(**updated_content)
+        updated_sale = Sale.objects.get(id=id)
+        return JsonResponse(
+            updated_sale,
+            encoder=SaleEncoder,
+            safe=False,
+        )
+    else:  # DELETE
+        count, _ = Sale.objects.filter(id=id).delete()
         return JsonResponse({"deleted": count > 0})
